them add property, property form component
